+++ custom_components/silla_prism/number.py
@@ -8,7 +8,6 @@ from homeassistant.components.number import (
     NumberDeviceClass,
     NumberEntity,
     NumberEntityDescription,
-    NumberMode,
 )
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import EntityCategory
@@ -21,9 +20,6 @@ from .entity import PrismBaseEntity
 from .entry_data import RuntimeEntryData
 
 _LOGGER = logging.getLogger(__name__)
-
-MIN_CHARGE_CURRENT = 6
-SOLAR_BALANCE_MANUAL_CURRENT_OVERRIDE = "solar_balance_manual_current_override"
 
 
 async def async_setup_entry(
@@ -162,19 +158,6 @@ class PrismNumber(PrismBaseEntity, NumberEntity):
         #     value,
         # )
         int_value = int(value)
-        if self.entity_description.translation_key == SOLAR_BALANCE_MANUAL_CURRENT_OVERRIDE:
-            if int_value >= MIN_CHARGE_CURRENT:
-                self._entry_data.solar_balance_manual_current_overrides[
-                    self._port
-                ] = int_value
-            else:
-                self._entry_data.solar_balance_manual_current_overrides.pop(
-                    self._port, None
-                )
-            self._attr_native_value = int_value
-            self.async_write_ha_state()
-            return
-
         await mqtt.async_publish(self.hass, self._topic_out, int_value)
 
 
@@ -201,18 +184,5 @@ NUMBERS: tuple[PrismNumberEntityDescription, ...] = (
         native_max_value=16,
         has_entity_name=True,
         translation_key="set_current_limit",
-    ),
-    PrismNumberEntityDescription(
-        key="solar_balance_manual_current_override_{}",
-        topic="",
-        topic_out=None,
-        entity_category=EntityCategory.CONFIG,
-        device_class=NumberDeviceClass.CURRENT,
-        native_min_value=0,
-        native_max_value=16,
-        native_step=1,
-        mode=NumberMode.SLIDER,
-        has_entity_name=True,
-        translation_key=SOLAR_BALANCE_MANUAL_CURRENT_OVERRIDE,
     ),
 )